+++ Source/Form/CustomizableLabel.swift
@@ -7,7 +7,12 @@ public protocol CustomizableTitleLabel {
     var titleTextColor: UIColor { get set }
 }
 
-typealias CustomizableLabels = CustomizableTitleLabel & CustomizableDetailLabel
+typealias CustomizableLabel = CustomizableTitleLabel & CustomizableDetailLabel
+
+// It's inconsistent using plural for this, and singular everywhere else.
+// Will be marked `unavailable` in the future.
+@available(iOS, deprecated, renamed: "CustomizableLabel")
+typealias CustomizableLabels = CustomizableLabel
 
 public extension CustomizableTitleLabel where Self: FormItem {