+++ Source/FormItems/PrecisionSliderFormItem.swift
@@ -15,7 +15,7 @@ import UIKit
 Behind the scenes this creates a `PrecisionSlider`. This is not a standard Apple control.
 Please contact Simon Strandgaard if you have questions regarding it.
 */
-public class PrecisionSliderFormItem: FormItem, CustomizableLabels {
+public class PrecisionSliderFormItem: FormItem, CustomizableLabel {
     
 	override func accept(visitor: FormItemVisitor) {
 		visitor.visit(object: self)