+++ Source/FormItems/TextFieldFormItem.swift
@@ -1,7 +1,7 @@
 // MIT license. Copyright (c) 2021 SwiftyFORM. All rights reserved.
 import UIKit
 
-public class TextFieldFormItem: FormItem, CustomizableLabels {
+public class TextFieldFormItem: FormItem, CustomizableLabel {
     
 	override func accept(visitor: FormItemVisitor) {
 		visitor.visit(object: self)