+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Contributions by Luca Pizzini. Thank you.
    The background color is now customizable for buttons.
    Extended the README with guidance for how to install SwiftyFORM manually.
+ * Renamed from `CustomizableLabels` to `CustomizableLabel`, so it's consistent with the new `CustomizableButton` protocol.
 
 ### 1.8.5